move typography stories to stories directory
